Move chat plugins to their own directory
